requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sqlalchemy==2.0.30
python-dotenv==1.1.0
email-validator==2.1.2

